Added positive-negative nvtx distributions for ngen, pu weights
Added number_of_vertices distributions (puli-up)
for the negative and positive case of ngen weight
(the one for the weird NLO accounting).

At the moment the overall weight = ngen * pu-weight
where ngen is 1 or -1 depending on some "ngen" parameter.
The pile-up reweighting is the distribution in template_cfg.py file
-- the one derived for 74v datasets.