Added HLT efficiencies
But EGamma POG has no distr yet -- setting eff = 1.

And I don't get the HLT object fired the trigger,
the eff-s are calculated according to the selected candidates
(in principle, it's mostly the trigger objects).

The dilepton case is managed with
 1 - (1 - eff_l1)(1 - eff_l2)
-- eff_l1, eff_l2 are efficiencies of two leptons.

TODO: check if TH2F is accessed in correct order (pt/eta or eta/pt)
and works fine with outside-range values.